@@ -2,7 +2,7 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.8.21'
+        kotlin_version = '1.9.10'
     }
     repositories {
         google()
@@ -10,11 +10,11 @@ buildscript {
         gradlePluginPortal()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:8.0.1'
+        classpath 'com.android.tools.build:gradle:8.1.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
 plugins {
-    id 'com.google.devtools.ksp' version '1.8.21-1.0.11' apply false
+    id 'com.google.devtools.ksp' version '1.9.10-1.0.13' apply false
 }
