@@ -22,12 +22,10 @@ import de.luhmer.owncloudnewsreader.helper.ThemeChooser;
 import de.luhmer.owncloudnewsreader.model.OcsUser;
 import de.luhmer.owncloudnewsreader.ssl.MemorizingTrustManager;
 
-import static com.nextcloud.android.sso.Constants.ACCOUNT_TYPE_PROD;
-
 public class TestApiModule extends ApiModule {
 
     private static final String TAG = TestApiModule.class.getCanonicalName();
-    private Application application;
+    private final Application application;
 
     public static String DUMMY_ACCOUNT_AccountName = "test-account";
     public static String DUMMY_ACCOUNT_username = "david";
@@ -46,7 +44,7 @@ public class TestApiModule extends ApiModule {
         SingleSignOnAccount ssoAccount = new SingleSignOnAccount(
                 DUMMY_ACCOUNT_AccountName, DUMMY_ACCOUNT_username,
                 DUMMY_ACCOUNT_token, DUMMY_ACCOUNT_server_url,
-                ACCOUNT_TYPE_PROD
+                "prod"
         );
 
         OcsUser userInfo = new OcsUser("1", DUMMY_ACCOUNT_username);
