@@ -23,8 +23,6 @@ package de.luhmer.owncloudnewsreader;
 
 import android.app.Activity;
 import android.app.Dialog;
-import android.app.DialogFragment;
-import android.content.DialogInterface;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
@@ -34,10 +32,12 @@ import android.view.WindowManager.LayoutParams;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AlertDialog;
+import androidx.fragment.app.DialogFragment;
+
 import java.io.IOException;
 import java.util.Formatter;
 
-import androidx.appcompat.app.AlertDialog;
 import de.luhmer.owncloudnewsreader.async_tasks.DownloadChangelogTask;
 import de.luhmer.owncloudnewsreader.view.ChangeLogFileListView;
 
