@@ -4,6 +4,8 @@ import android.content.Context;
 import android.content.SharedPreferences;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import de.luhmer.owncloudnewsreader.SettingsActivity;
@@ -15,7 +17,8 @@ public class Search {
 
     private static final String SEARCH_IN_TITLE = "0";
     private static final String SEARCH_IN_BODY = "1";
-    
+    private static final String SEARCH_IN_BOTH = "2";
+
     public static List<RssItem> PerformSearch(Context context, Long idFolder, Long idFeed, String searchString, SharedPreferences mPrefs) {
         DatabaseConnectionOrm.SORT_DIRECTION sortDirection = DatabaseUtilsKt.getSortDirectionFromSettings(mPrefs);
         DatabaseConnectionOrm dbConn = new DatabaseConnectionOrm(context);
@@ -43,11 +46,13 @@ public class Search {
                                        final DatabaseConnectionOrm dbConn,
                                        final SharedPreferences mPrefs) {
         String sql = "";
-        String searchIn = mPrefs.getString(SettingsActivity.SP_SEARCH_IN,"0");
+        String searchIn = mPrefs.getString(SettingsActivity.SP_SEARCH_IN, SEARCH_IN_BOTH); 
         if(searchIn.equals(SEARCH_IN_TITLE)) {
             sql = dbConn.getAllItemsIdsForFeedSQLFilteredByTitle(idFeed, false, false, sortDirection, searchString);
         } else if(searchIn.equals(SEARCH_IN_BODY)) {
             sql = dbConn.getAllItemsIdsForFeedSQLFilteredByBodySQL(idFeed, false, false, sortDirection, searchString);
+        } else if (searchIn.equals(SEARCH_IN_BOTH)) {
+            sql = dbConn.getAllItemsIdsForFeedSQLFilteredByTitleAndBodySQL(idFeed, false, false, sortDirection, searchString);
         }
         return sql;
     }
@@ -58,11 +63,14 @@ public class Search {
                                          final DatabaseConnectionOrm dbConn,
                                          final SharedPreferences mPrefs) {
         String sql = "";
-        String searchIn = mPrefs.getString(SettingsActivity.SP_SEARCH_IN,"0");
+        String searchIn = mPrefs.getString(SettingsActivity.SP_SEARCH_IN, SEARCH_IN_BOTH);
         if(searchIn.equals(SEARCH_IN_TITLE)) {
-            sql = dbConn.getAllItemsIdsForFolderSQLSearch(ID_FOLDER, sortDirection, RssItemDao.Properties.Title.columnName, searchString);
+            sql = dbConn.getAllItemsIdsForFolderSQLSearch(ID_FOLDER, sortDirection, Collections.singletonList(RssItemDao.Properties.Title.columnName), searchString);
         } else if(searchIn.equals(SEARCH_IN_BODY)) {
-            sql = dbConn.getAllItemsIdsForFolderSQLSearch(ID_FOLDER, sortDirection, RssItemDao.Properties.Body.columnName, searchString);
+            sql = dbConn.getAllItemsIdsForFolderSQLSearch(ID_FOLDER, sortDirection, Collections.singletonList(RssItemDao.Properties.Body.columnName), searchString);
+        } else if(searchIn.equals(SEARCH_IN_BOTH)) {
+            var columns = Arrays.asList(RssItemDao.Properties.Body.columnName, RssItemDao.Properties.Title.columnName);
+            sql = dbConn.getAllItemsIdsForFolderSQLSearch(ID_FOLDER, sortDirection, columns, searchString);
         }
 
         return sql;
